feat: Add AWS provider and account module
This commit adds the AWS provider with two profiles, a new account provider, and an account module. It also includes a simple resource to test the new account provider and outputs for the created account's ID and ARN. The required Terraform version is set to 1.0 or higher, and the minimum required version of the AWS provider is 5.0 or higher.